Build package on update-pkg action
We never really did this before. I'm not sure how it changes the
deployments. What might've happened is that the package did not get
built, but the cache is updated, causing the derivation to be missing.

Ideally, this should not happen. Nix should've seen that the derivation
is not in /nix/store and should've built it during `terraform apply`.
I'm not sure why it's not doing this either, but this fix should work.

We also want the update-pkg to fail early if the package cannot be built
anyway. It makes more sense that way.