Make Deno builds reproducible
DENO_DIR is weird. They actually throw the current timestamp in DENO_DIR
for no particular reason. This ensures that builds are NEVER
reproducible! What a wonderful idea.

This commit creates a deno-cache derivation similar to go-modules and
filters the .metadata.json file so that it's reproducible. This should
fix all the build inconsistencies.